quick sort algorithm

+++ sort.h
@@ -27,7 +27,7 @@ void bubble_sort(int *array, size_t size);
 void insertion_sort_list(listint_t **list);
 void selection_sort(int *array, size_t size);
 void quick_sort(int *array, size_t size);
-size_t partition(int *array, size_t lb, size_t ub, size_t size);
+int partition(int *array, int lb, int ub, int size);
 void sort(int *array, size_t lb, size_t ub, size_t size);
 void shell_sort(int *array, size_t size);
